Address evaluator Sonar maintainability gate
